PS1Request size parameter is Int

+++ Sources/SwiftMAST/Request+json model/MASTRequest.swift
@@ -130,10 +130,10 @@ public struct PS1Request {
     private let downloadFileUrl = ""
     private let fitsCutRequestUrl = "http://ps1images.stsci.edu/cgi-bin/fitscut.cgi"
     private(set) var parameters:[String: Any]
-    private let size:Float
+    private let size:Int
     private let format:String
     
-    public init(ra: Float, dec: Float, size: Float = 4000, filters:PS1Filter = .grizy, format: PS1FileType = .fits, imageTypes: [PS1ImageType] = [.stack]) {
+    public init(ra: Float, dec: Float, size: Int = 4000, filters:PS1Filter = .grizy, format: PS1FileType = .fits, imageTypes: [PS1ImageType] = [.stack]) {
         self.parameters = [:]
         parameters["filters"] = filters.id as Any
         parameters["type"] = imageTypes.map{$0.id}.joined(separator: ",") as Any
